pyproject.toml
[project]
name = "service-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.3"
dependencies = [
    "cryptography>=45.0.5",
    "dotenv>=0.9.9",
    "google-adk>=1.8.0",
    "litellm>=1.74.8",
]
